+++ backend/src/main/java/ch/time2log/backend/domain/models/ActivitySummary.java
@@ -1,9 +1,10 @@
 package ch.time2log.backend.domain.models;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record ActivitySummary(
         UUID activityId,
         String activityName,
-        int totalHours
+        BigDecimal totalHours
 ) {}
